Refactor cnv_data_extraction.py for Improved Data Processing
Added missing import statement for 'pandas'
Modifed 'extract_data_from_cnvkit_source' function to read data into a Dataframe and format it correctly.
Created a tuple 'output' in the required format to plot on a graph of 'cnv_data'.
Printed the formatted output.

These changes improve the data extraction and formatting process in the 'cnv_data_extraction.py' script, ensuring the output is correclty structured to be plotted onto the chromosome graph